pyproject.toml
[project]
name = "trading-automation-fullstack"
version = "0.1.0"
description = "Professional trading automation system"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "requests>=2.31.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
    "yfinance>=1.0",
    "scipy>=1.13.1",
    "lxml>=6.0.2",
    "html5lib>=1.1",
    "joblib>=1.5.3",
    "scikit-learn>=1.6.1",
    "alpaca-py>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "moto>=5.1.19",
]
